+++ HISTORY.md
@@ -4,6 +4,23 @@
 
 ## [unreleased]
 
+## [0.4.0] - 2026-02-05 | Event logging (moved from DiscordConnect)
+
+- feat: event logging implemented (from DiscordConnect)
+  - all events can be enabled or disabled:
+    - player deaths
+    - player connect / disconnect
+    - player removes object
+    - player destroys object
+    - npc death by non player
+    - mount death by player
+    - (all) animal death by player
+    - season changes
+    - weather changes
+    - player teleport events
+  - (optional) logging to discord by setting the channelId
+    - can be configured for each event
+
 ## [0.3.0] - 2026-02-04 | Custom Mount Names, Sleep-Announcement-Feature
 
 - feat: moved sleep announcement from tools to this plugin